feat: changed deps structure
